tau2-examples/pyproject.toml
[project]
name = "tau2-examples"
version = "0.0.1"
description = "Tau2 benchmark examples with policy enforcement"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tau2",
    "sasy",
    "python-dotenv>=1.0.0",
    "google-cloud-aiplatform>=1.132.0",
]

[project.scripts]
tau2-instrumented = "tau2_examples.cli:run"
tau2-instrumented-llm-judge = "tau2_examples.cli_llm_policy:run"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["tau2_examples"]

[tool.uv.sources]
sasy = { workspace = true }
tau2 = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
